LayoutTests/media/unsupported-rtsp.html

<!DOCTYPE html>
<title>Test that QuickTime file with RTSP URL generates a load error.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = "rtsp://a2047.v1411b.c1411.g.vq.akamaistream.net/5/2047/1411/2_h264_650/1a1a1ae454c430950065de4cbb2f94c226950c7ae655b61a48a91475e243acda3dac194879adde0f/wwdc_2006_2_650.mov";
    video.onerror = t.step_func_done(function () {
        assert_equals(video.error.code, MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);
        assert_equals(video.networkState, HTMLMediaElement.NETWORK_NO_SOURCE);
    });
});
</script>
