LayoutTests/media/unsupported-rtsp-expected.txt

Test that QuickTime file with RTSP URL generates a load error.

EVENT(error)
TEST(video.error) OK
TEST(video.error.code = MediaError.MEDIA_ERR_NETWORK) OK
END OF TEST
